Add container assignment to harvesters
For each "sourceContainer" in room memory, map that container to a source
and to a creep. That creep will then simply harvest from that source, and
put the energy in that container. Ideally this creep will sit in place and
simply continuously transfer energy from the source to the container that
is in the same position or right next to it.

If no sourceContainers exist, fall back to legacy behavior.