Introduce role for newSpawnBuilder
After a neutral controller has been claimed, the claimer will drop a
NewSpawnFlag on the ClaimFlag spot. This will trigger any existing spawn
that has capacity to produce a newSpawnBuilder creep, who will move to the
NewSpawnFlag, create a construction site for a spawn, and then start
trying to build it using nearby sources.